read() also takes config
